Reorganizing functions in coyote_simple: now it behaves the same if the received frame are ERR, RPY, MSG or ANS.
Coyote_simple will not manage the message buffer if frame_available_cb is
defined, and will refuse messages if frame_available_cb and
complete_message_cb are undefined.

When it manages the message buffer, there are some limits: 2KB for MSG and
ERR, and 10KB for RPY and ANS. If the message exceeds this length, it will
truncated.