idls: * removing --merge-option from Makefile.am to avoid overwriting source code while developing with af-gen tool.
af-gen:
* making af-gen to detect when a module attribute has the same name as the
module holding it, showing an error and then giving up.

* fixing an non existent reference to enum types defined at C layer from
the main header.

* adding support to defined several relation between the same modules. In order
to support this feature it has been added a new construct: "bind as" which allows
to configure the new to be used on multiple relations. This new sentence tweaks
the services generated and attributes defined implicitly.

* making af-gen to detect and show an error if several relation are defined between
the same modules without using "bind as" sentence.

* adding support to af-gen IDL language to also document modules using short and 
description document blocks. Refactoring source code inside the af_gen_doc_services
module.

* making af-gen to not generate source test modules by default. Added a new option
to reenable it: --enable-test