libcoyote: * making coyote_xml_destroy_message_full to nullify the message deallocated
libvortex:
* implemented RFC3080 required to automatically asign odd channel numbers for those
connections in initiator role and even numbers for those acting as listener role.

* adding support for an extended start message reception. now it is posible 
to register an extended version for the start handler which allows to get access
to all available options received on channel creation request. previous code writing
with the normal start handler keeps running.  This is implemented throught a new
function: 

   vortex_profiles_register_extended_start
which allows to register an extended start, which overrides the start handler defined.
Added more documentation explaining how to use the start handlers and how they interact.


* modified vortex_profile_is_start_defined to return if an normal start or extended one
is already defined.

* modified vortex_profile_invoke_start to invoke the extended start defined or the normal
start handler if the previous wasn't being defined.