libvortex: * adding support to vortex client tool to be able to support enabling TLS profile support.
* adding support to notify when a connection have being closed through a 
VortexConnectionOnClose handler. This allows to perform some especific 
operations over connection closed. Connection Close is executed just
before its resouces are deallocated. Added API to allow setting the 
on close handler:
  - vortex_connection_set_on_close [public]

* making vortex_connection_close_all_channels to be public so profile
implementators could use it.

* adding support for implicit piggyback support. Now, piggyback message
sent at the start message or on its reply, is notified to the user space
as a received frame. This allows to support with the same source code
replies with piggyback or withtout it. Added API:
  - vortex_channel_have_piggyback [public]
  - vortex_channel_get_piggyback  [public]
  - vortex_channel_set_piggyback  [private]

* added a new top level mechanism to receive frames in a synchroumos manner
without using wait reply which is more especific and expensive. This 
method is based on two functions:
  - vortex_channel_queue_reply [public]: which just queue any frame received
    into a user space queue.
  - vortex_channel_get_reply [public]: which fetch the next frame recevied
    on the selected queue, including support to fetch the optional piggyback
    received.

* added new function to create VortexFrame instances from user space data:
  - vortex_frame_create [public]

* added initial client TLS implementation proposal, still not fully working,
contained at vortex_tls.[hc].