libvortex:
* [fix] starting the ansification process, at this moment only
comments are translated into the /* ansi form */ from the // moderm
form.

* [fix] fixing a really incredible bug based on a race condition
between building the next frame to be sent and queueing the rest of
the message due to remote buffer size is complete but, under some
circumstances the frame generated is fast enough to reach remote side,
be delivered into the application space, generate a new SEQ frame, and
make local vortex reader to start a resequencing process but the
initial !!already didn't queue the rest of the message to be
resequenced.!!!! Reorganized vortex sequencer code to avoid previous
race condition: threads lot of fun!

* [fix] fixing initial values used for the max seq no that the remote
peer is willing to accept and the local buffer for each channel from
4096 to 4095, which represent a total octect counting of 4096 starting
from 0 up to 4095. Making vortex reader, vortex channel and vortex
sequencer to reorganize its operational calculus to this change. 

* [fix] fixing vortex_channel_update_incoming_buffer to not report
vortex reader to send a SEQ frame if it is detected that current
configuration will not allow this. Making this function to directly
report current ackno and window size values so the vortex reader
completely depends on this function.

* [new] adding support to vortex frame module to generate an unique
global identifier to uniquely identify every frame, mainly to allow
memory debuging but also providing aplication space a way to identify
frames. API added:

    - vortex_frame_get_id 

* [fix] making vortex_frame_create to use services provided by
vortex_frame_create_full to build new frames. This new function not
only uses memcpy instructions to build new frames but also adds
support to specify frame content type and content transfer encoding
values. API added:
    
    - vortex_frame_create_full