af-arch: * [fix] fixing some coments at "what is the af-arch" document. * [fix] updating modules-version.txt file to upgrade officially vortex library release to 0.8.9.
libvortex:
* [fix] making vortex greetings to update message numbers to be expected,
rather than making an exception for the channel 0. 

* [fix] making vortex_channel_empty_new to set as default reply sent and
reply received for channel 0 values {1, 1}. This solves the problem found
for the message 0 that is never sent for the channel 0.

* [fix] fixing a bug while processing start messages. If the xml received contains
garbage, that is spaces, tabulars, returns, between the <start> element and the <profile>
element, the libxml library set as a child node to the <start> element the garbarge!!
Added some additional checks to test if the node read is a profile node or something
else.

* [fix] making vortex reader to accept first messages that starts from with msgno=1.
API added to control this function:

  - vortex_reader_allow_msgno_starting_from_1.

This will make vortex library to interop with beepcore-java library.